NXP-20152: add spherical coords to render filenames; update converter and testes

+++ nuxeo-platform-3d-convert/src/main/resources/scripts/pipeline.py
@@ -10,7 +10,7 @@ def sphericalCoords(s):
     try:
         if s == "":
             return 0, 0
-        zenith, azimuth = map(float, s.split(','))
+        zenith, azimuth = map(int, s.split(','))
         return zenith, azimuth
     except:
         raise argparse.ArgumentTypeError("Coordinates must be zenith,azimuth")
